move prisma singleton
