+++ portfolio-next/pages/api/getStations.ts
@@ -1,5 +1,5 @@
 import type {NextApiRequest, NextApiResponse} from 'next';
-import prisma from '../db' 
+import prisma from './db' 
 
 export default async (req: NextApiRequest, res: NextApiResponse) => {
     if( req.method !== 'GET'){
