+++ portfolio-next/pages/api/ridesFromStation.ts
@@ -1,5 +1,5 @@
 import type {NextApiRequest, NextApiResponse} from 'next';
-import prisma from '../db' 
+import prisma from './db' 
 
 export default async (req: NextApiRequest, res: NextApiResponse) => {
     const query = req.query;
